+++ busy_beaver/adapters/utilities.py
