+++ tests/adapters/cassettes/test_all_user_stars.yaml
