+++ tests/adapters/utilities_test.py
@@ -1,6 +1,10 @@
 import pytest
 
-from adapters.utilities import create_github_navigation_panel, APINav, page_from_url
+from busy_beaver.adapters.utilities import (
+    APINav,
+    create_github_navigation_panel,
+    page_from_url,
+)
 
 
 @pytest.mark.unit
@@ -37,11 +41,14 @@ def test_parsing_header_links_empty():
         create_github_navigation_panel(link_header)
 
 
-@pytest.mark.parametrize("url, expected_page_num", [
-    ("https://api.github.com/user/4369343/events/public?page=1", 1),
-    ("https://api.github.com/user/4369343/events/public?page=3", 3),
-    ("https://api.github.com/user/4369343/events/public?page=10", 10),
-])
+@pytest.mark.parametrize(
+    "url, expected_page_num",
+    [
+        ("https://api.github.com/user/4369343/events/public?page=1", 1),
+        ("https://api.github.com/user/4369343/events/public?page=3", 3),
+        ("https://api.github.com/user/4369343/events/public?page=10", 10),
+    ],
+)
 def test_page_from_url(url, expected_page_num):
     # Act
     result = page_from_url(url)
